Trying to address no STARTTLS offerings
As noted in #1536 if the server side doesn't show STARTTLS
it should exit and label it accordingly.

This is an attempt to address that. It does that by using an early
exit using fatal().

For this to achieve the starttls*dialog() functions were changed
so that a return value from starttls_full_read() of 3,
signalling the STARTTLS pattern wasn't found is passed back to
the parent fd_socket().

Furthermore:
  * starttls*dialog() functions' debug statements were moved into
    starttls_full_read() / starttls_just_send()
  * starttls_full_read() + starttls_just_send() were improved for
    readability and debugging
  * minor bugs were squashed (e.g. fd_socket return values =!0 always
    were referring to STARTTLS also when no STARTTLS was requested)

This was tested (negative + test and positive) for FTP and SMTP
which worked as expected. For IMAP this needs to be done. The
POP3 server side doesn't seem to advertise STARTTLS (TBC), for
NNTP I haven't bothered to look at it.

XMPP and postgreSQL use starttls_io() which need to be changed
as well.