Fix typo in Strong grade Ciphers AEAD docs
Thanks for an awesome tool.
<li><code>Strong grade Ciphers</code> (AEAD): 'AESGCM:CHACHA20:AESGCM:CamelliaGCM:AESCCM8:AESCCM'</li>
The Strong AEAD ciphers on line 243 above don't make sense. I suspect a copy paste bug. The 2nd AESGCM should be ARIAGCM no?
Sorry I haven't checked the code to see if the same issue is in the code or if this was just a documentation glitch.